requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
drf-yasg==1.21.15
sqlparse==0.5.5
